Simplify Configuration System and Improve Light Creation Performance
This PR refactors the Stoplight configuration system to eliminate the `ConfigProvider` layer
and implement a more direct, prototype-based configuration approach. The changes result
in **2.26x faster light creation** while maintaining full backward compatibility.

- **2.26x faster light creation** (398k → 900k operations/second)
- Eliminated unnecessary object allocations during configuration resolution
- Simple configuration inheritance using prototype pattern
- Simplified the main `Stoplight()` function to use direct light creation

**Enhanced `Light::Config`:**
```ruby
config = Stoplight.config_provider.provide(name, settings)

config = Stoplight.default_config.with(name: name, **settings)
```

**New light creation methods:**

```ruby
Stoplight.light(name, **settings)

Stoplight.system_light(name, **settings)
```

- **Self-validating configs**: `Light::Config#validate_config!` ensures compatibility
- **Immutable configuration objects**: All config changes create new instances
- **Better encapsulation**: Configuration logic contained within config objects

✅ **Fully backward compatible** - All existing APIs work unchanged:
- `Stoplight(name, **settings)`
- `Stoplight.configure { |config| ... }`
- `light.with(**new_settings)`
- All configuration options and validation
