+++ lib/stoplight/config/system_config.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Stoplight
+  module Config
+    SystemConfig = LibraryDefaultConfig
+  end
+end
