+++ spec/stoplight/data_store/redis_spec.rb
@@ -4,7 +4,7 @@ require "connection_pool"
 require "spec_helper"
 
 RSpec.describe Stoplight::DataStore::Redis, :redis do
-  let(:config) { Stoplight.config_provider.provide(name, window_size:, cool_off_time:) }
+  let(:config) { Stoplight.default_config.with(name:, window_size:, cool_off_time:) }
   let(:name) { ("a".."z").to_a.shuffle.join }
   let(:failure) { Stoplight::Failure.new("class", "message", Time.new - 60) }
   let(:other) { Stoplight::Failure.new("class", "message 2", Time.new) }
@@ -146,7 +146,7 @@ RSpec.describe Stoplight::DataStore::Redis, :redis do
 
     it_behaves_like "data store metrics" do
       context "when JSON is invalid" do
-        let(:config) { Stoplight.config_provider.provide(name, error_notifier: ->(_error) {}) }
+        let(:config) { Stoplight.default_config.with(name:, error_notifier: ->(_error) {}) }
 
         it "handles it without an error" do
           expect(failure).to receive(:to_json).and_return("invalid JSON")
