+++ spec/stoplight/light/yellow_run_strategy_spec.rb
@@ -6,10 +6,11 @@ RSpec.describe Stoplight::Light::YellowRunStrategy do
   subject(:strategy) { described_class.new(config) }
 
   let(:config) do
-    Stoplight.config_provider.provide("foo",
+    Stoplight.default_config.with(
       data_store:,
       traffic_recovery:,
-      notifiers: [notifier])
+      notifiers: [notifier]
+    )
   end
   let(:notifier) { instance_double(Stoplight::Notifier::Base) }
   let(:traffic_recovery) { instance_double(Stoplight::TrafficRecovery::Base) }
